add feature gate
