+++ demo/caller/Cargo.toml
@@ -6,4 +6,4 @@ edition = "2018"
 publish = false
 
 [dependencies]
-wa-demo = { path = "../wa" }
+wa-demo = { path = "../wa", features = ["proc-macro-server"] }
