+++ demo/wa/Cargo.toml
@@ -10,3 +10,7 @@ proc-macro = true
 
 [dependencies]
 watt = "0.4"
+
+[features]
+default = []
+proc-macro-server = ["watt/proc-macro-server"]
